chore(cliff): improve template
There was no empty line between the list of new contributors and the
next release. A previous attempt to fix this resulted in 2 empty lines
when there were no new contributors.
There should also be an empty line before the footer.
This change takes care of both of these cases.